oidc plugin cert population
